test(test): add tests for `ts-node` related dependencies is not found
