+++ packages/test/src/utils/index.ts
@@ -1,1 +1,2 @@
 export { default as mockConsole } from './mockConsole';
+export { default as loadTSConfigFile } from './loadTsConfig';
